+++ .gitignore
@@ -1,3 +1,4 @@
+.DS_Store
 npm-debug.log*
 node_modules
 .npm
