+++ gulp/worker/compless.js
@@ -45,21 +45,37 @@ gulp.task('compless:i2c', ()=> {
     .pipe(gulp.dest(config.rootDirs.dist));
 });
 
-gulp.task('compless:worker', ()=> {
-  return gulp.src(config.paths.script.worker.src)
-    .pipe(concat('worker.js'))
+gulp.task('compless:worker_gpio', ()=> {
+  return gulp.src(config.paths.script.worker.gpio.src)
+    .pipe(concat('worker.gpio.js'))
     .pipe(insert.wrap('(function(){', '})()'))
+    .pipe(gulp.dest(config.rootDirs.dist))
+    .pipe(rename({
+      suffix: '.min',
+    }))
+    .pipe(babel({
+      presets: [],
+      plugins: ['transform-es2015-arrow-functions'],
+    }))
+    .pipe(uglify({
+      compress:true,
+    }))
     .pipe(gulp.dest(config.rootDirs.dist));
-
-  // .pipe(rename({
-  //   suffix: '.min',
-  // }))
-  // .pipe(babel({
-  //   presets: [],
-  //   plugins: ['transform-es2015-arrow-functions'],
-  // }))
-  // .pipe(uglify({
-  //   compress:true,
-  // }))
-  // .pipe(gulp.dest(config.rootDirs.dist));
+});
+gulp.task('compless:worker_i2c', ()=> {
+  return gulp.src(config.paths.script.worker.i2c.src)
+    .pipe(concat('worker.i2c.js'))
+    .pipe(insert.wrap('(function(){', '})()'))
+    .pipe(gulp.dest(config.rootDirs.dist))
+    .pipe(rename({
+      suffix: '.min',
+    }))
+    .pipe(babel({
+      presets: [],
+      plugins: ['transform-es2015-arrow-functions'],
+    }))
+    .pipe(uglify({
+      compress:true,
+    }))
+    .pipe(gulp.dest(config.rootDirs.dist));
 });
